v5.3 updated code

+++ app/Http/Controllers/Admin/CustomRoleController.php
@@ -19,9 +19,11 @@ class CustomRoleController extends Controller
     public function store(Request $request)
     {
         $request->validate([
-            'name' => 'required|unique:admin_roles',
+            'name' => 'required|unique:admin_roles|max:191',
+            'modules'=>'required|array|min:1'
         ],[
-            'name.required'=>'Role name is required!'
+            'name.required'=>trans('messages.Role name is required!'),
+            'modules.required'=>trans('messages.Please select atleast one module')
         ]);
         DB::table('admin_roles')->insert([
             'name'=>$request->name,
@@ -37,16 +39,26 @@ class CustomRoleController extends Controller
 
     public function edit($id)
     {
+        if($id == 1)
+        {
+            return view('errors.404');
+        }
         $role=AdminRole::where(['id'=>$id])->first(['id','name','modules']);
         return view('admin-views.custom-role.edit',compact('role'));
     }
 
     public function update(Request $request,$id)
     {
+        if($id == 1)
+        {
+            return view('errors.404');
+        }
         $request->validate([
-            'name' => 'required|unique:admin_roles,name,'.$id,
+            'name' => 'required|max:191|unique:admin_roles,name,'.$id,
+            'modules'=>'required|array|min:1'
         ],[
-            'name.required'=>'Role name is required!'
+            'name.required'=>trans('messages.Role name is required!'),
+            'modules.required'=>trans('messages.Please select atleast one module')
         ]);
 
         DB::table('admin_roles')->where(['id'=>$id])->update([
@@ -61,6 +73,10 @@ class CustomRoleController extends Controller
     }
     public function distroy($id)
     {
+        if($id == 1)
+        {
+            return view('errors.404');
+        }
         $role=AdminRole::where(['id'=>$id])->delete();
         Toastr::success(trans('messages.role_deleted_successfully'));
         return back();
@@ -68,12 +84,12 @@ class CustomRoleController extends Controller
 
     public function search(Request $request){
         $key = explode(' ', $request['search']);
-        $rl=AdminRole::
-        where(function ($q) use ($key) {
+        $rl=AdminRole::where('id','!=','1')
+        ->where(function ($q) use ($key) {
             foreach ($key as $value) {
                 $q->orWhere('name', 'like', "%{$value}%");
             }
-        })->limit(50)->get();
+        })->latest()->limit(50)->get();
         return response()->json([
             'view'=>view('admin-views.custom-role.partials._table',compact('rl'))->render(),
             'count'=>$rl->count()
